Move Select Map Terrain to its own file
